rearranged code for visual continuity
